Extract shared serving_chat test helpers into tests/helpers/
