+++ tests/entrypoints/test_stream_finish_reason.py
@@ -16,30 +16,20 @@ Key invariants tested:
   - voice/speaker parameter compatibility in chat completions
 """
 
-import enum
 import json
 from unittest.mock import MagicMock
 
 import pytest
-
-# Python 3.10 compat: StrEnum was added in 3.11
-if not hasattr(enum, "StrEnum"):
-
-    class _StrEnum(str, enum.Enum):
-        """Minimal StrEnum backport for Python 3.10."""
-
-    enum.StrEnum = _StrEnum  # type: ignore[attr-defined]
-
 from vllm.entrypoints.openai.chat_completion.protocol import (
     ChatCompletionRequest,
-    ChatCompletionResponseStreamChoice,
 )
-from vllm.entrypoints.openai.engine.protocol import DeltaMessage
-from vllm.entrypoints.openai.models.serving import OpenAIServingModels
-from vllm.outputs import CompletionOutput, RequestOutput
 
-from vllm_omni.entrypoints.openai.serving_chat import OmniOpenAIServingChat
-from vllm_omni.outputs import OmniRequestOutput
+from tests.helpers.serving_chat import (
+    build_serving_chat,
+    make_audio_omni_output,
+    make_request,
+    make_text_omni_output,
+)
 
 pytestmark = [pytest.mark.core_model, pytest.mark.cpu]
 
@@ -47,126 +37,6 @@ pytestmark = [pytest.mark.core_model, pytest.mark.cpu]
 # ---------------------------------------------------------------------------
 # Helpers
 # ---------------------------------------------------------------------------
-
-
-def _make_text_omni_output(
-    request_id: str = "test-req",
-    text: str = "hello",
-    token_ids: list[int] | None = None,
-    finish_reason: str | None = None,
-    index: int = 0,
-    num_prompt_tokens: int = 3,
-) -> OmniRequestOutput:
-    """Build an OmniRequestOutput wrapping a text RequestOutput."""
-    if token_ids is None:
-        token_ids = [10, 11, 12]
-    res = RequestOutput(
-        request_id=request_id,
-        prompt="test",
-        prompt_token_ids=list(range(num_prompt_tokens)),
-        prompt_logprobs=None,
-        outputs=[
-            CompletionOutput(
-                index=index,
-                text=text,
-                token_ids=token_ids,
-                cumulative_logprob=0.0,
-                logprobs=None,
-                finish_reason=finish_reason,
-                stop_reason=None,
-            )
-        ],
-        finished=finish_reason is not None,
-    )
-    return OmniRequestOutput(
-        request_id=request_id,
-        final_output_type="text",
-        request_output=res,
-        finished=finish_reason is not None,
-    )
-
-
-def _make_audio_omni_output(
-    request_id: str = "test-req",
-    index: int = 0,
-    num_prompt_tokens: int = 3,
-) -> OmniRequestOutput:
-    """Build an OmniRequestOutput for audio (no torch dependency)."""
-    res = RequestOutput(
-        request_id=request_id,
-        prompt="test",
-        prompt_token_ids=list(range(num_prompt_tokens)),
-        prompt_logprobs=None,
-        outputs=[
-            CompletionOutput(
-                index=index,
-                text="",
-                token_ids=[],
-                cumulative_logprob=0.0,
-                logprobs=None,
-                finish_reason="stop",
-                stop_reason=None,
-            )
-        ],
-        finished=True,
-    )
-    return OmniRequestOutput(
-        request_id=request_id,
-        final_output_type="audio",
-        request_output=res,
-        finished=True,
-    )
-
-
-def _mock_audio_choices(index: int = 0, role: str = "assistant"):
-    return [
-        ChatCompletionResponseStreamChoice(
-            index=index,
-            delta=DeltaMessage(role=role, content="dGVzdA=="),
-            logprobs=None,
-            finish_reason="stop",
-        )
-    ]
-
-
-def _build_serving_chat():
-    """Create a minimal OmniOpenAIServingChat for testing."""
-    mock_engine = MagicMock()
-    mock_engine.errored = False
-
-    models = OpenAIServingModels(
-        engine_client=mock_engine,
-        base_model_paths=[],
-    )
-    mock_render = MagicMock()
-
-    instance = OmniOpenAIServingChat(
-        engine_client=mock_engine,
-        models=models,
-        response_role="assistant",
-        online_renderer=mock_render,
-        request_logger=None,
-        chat_template=None,
-        chat_template_content_format="auto",
-    )
-    instance._create_audio_choice = MagicMock(
-        side_effect=lambda omni_res, role, request, stream=False: _mock_audio_choices(
-            index=omni_res.request_output.outputs[0].index,
-            role=role,
-        )
-    )
-    return instance
-
-
-def _make_request(modalities: list[str], n: int = 1) -> ChatCompletionRequest:
-    req = ChatCompletionRequest(
-        model="test-model",
-        messages=[{"role": "user", "content": "hello"}],
-        n=n,
-        stream=True,
-    )
-    req.modalities = modalities  # type: ignore[attr-defined]
-    return req
 
 
 def _parse_sse_chunks(lines: list[str]) -> list[dict]:
@@ -202,12 +72,12 @@ async def _collect_stream(gen):
 @pytest.mark.asyncio
 async def test_single_modality_text_only_one_stop():
     """Text-only streaming: exactly one chunk has finish_reason='stop'."""
-    serving_chat = _build_serving_chat()
-    request = _make_request(modalities=["text"])
+    serving_chat = build_serving_chat()
+    request = make_request(modalities=["text"])
 
     async def result_generator():
-        yield _make_text_omni_output(text="he", token_ids=[10, 11], finish_reason=None)
-        yield _make_text_omni_output(text="llo", token_ids=[12], finish_reason="stop")
+        yield make_text_omni_output(text="he", token_ids=[10, 11], finish_reason=None)
+        yield make_text_omni_output(text="llo", token_ids=[12], finish_reason="stop")
 
     raw_lines = await _collect_stream(
         serving_chat.chat_completion_stream_generator(
@@ -233,13 +103,13 @@ async def test_single_modality_text_only_one_stop():
 @pytest.mark.asyncio
 async def test_multi_modal_text_audio_only_last_stop():
     """text+audio: text finish sends finish_reason=null, audio sends stop."""
-    serving_chat = _build_serving_chat()
-    request = _make_request(modalities=["text", "audio"])
+    serving_chat = build_serving_chat()
+    request = make_request(modalities=["text", "audio"])
 
     async def result_generator():
-        yield _make_text_omni_output(text="he", token_ids=[10, 11], finish_reason=None)
-        yield _make_text_omni_output(text="llo", token_ids=[12], finish_reason="stop")
-        yield _make_audio_omni_output()
+        yield make_text_omni_output(text="he", token_ids=[10, 11], finish_reason=None)
+        yield make_text_omni_output(text="llo", token_ids=[12], finish_reason="stop")
+        yield make_audio_omni_output()
 
     raw_lines = await _collect_stream(
         serving_chat.chat_completion_stream_generator(
@@ -269,16 +139,16 @@ async def test_multi_modal_text_audio_only_last_stop():
 @pytest.mark.asyncio
 async def test_multi_modal_n2_independent_per_choice():
     """n=2 with text+audio: each choice gets exactly one stop, at the end."""
-    serving_chat = _build_serving_chat()
-    request = _make_request(modalities=["text", "audio"], n=2)
+    serving_chat = build_serving_chat()
+    request = make_request(modalities=["text", "audio"], n=2)
 
     async def result_generator():
-        yield _make_text_omni_output(text="A", token_ids=[10], finish_reason=None, index=0)
-        yield _make_text_omni_output(text="B", token_ids=[20], finish_reason=None, index=1)
-        yield _make_text_omni_output(text="", token_ids=[11], finish_reason="stop", index=0)
-        yield _make_text_omni_output(text="", token_ids=[21], finish_reason="stop", index=1)
-        yield _make_audio_omni_output(index=0)
-        yield _make_audio_omni_output(index=1)
+        yield make_text_omni_output(text="A", token_ids=[10], finish_reason=None, index=0)
+        yield make_text_omni_output(text="B", token_ids=[20], finish_reason=None, index=1)
+        yield make_text_omni_output(text="", token_ids=[11], finish_reason="stop", index=0)
+        yield make_text_omni_output(text="", token_ids=[21], finish_reason="stop", index=1)
+        yield make_audio_omni_output(index=0)
+        yield make_audio_omni_output(index=1)
 
     raw_lines = await _collect_stream(
         serving_chat.chat_completion_stream_generator(
@@ -306,11 +176,11 @@ async def test_multi_modal_n2_independent_per_choice():
 @pytest.mark.asyncio
 async def test_single_modality_audio_only_one_stop():
     """Audio-only streaming: the audio chunk carries finish_reason='stop'."""
-    serving_chat = _build_serving_chat()
-    request = _make_request(modalities=["audio"])
+    serving_chat = build_serving_chat()
+    request = make_request(modalities=["audio"])
 
     async def result_generator():
-        yield _make_audio_omni_output()
+        yield make_audio_omni_output()
 
     raw_lines = await _collect_stream(
         serving_chat.chat_completion_stream_generator(
@@ -340,13 +210,13 @@ async def test_single_modality_audio_only_one_stop():
 async def test_declared_modality_not_produced_emits_fallback_stop():
     """If request.modalities declares ["text","audio"] but engine only produces
     text, a fallback stop chunk is emitted at stream end."""
-    serving_chat = _build_serving_chat()
-    request = _make_request(modalities=["text", "audio"])
+    serving_chat = build_serving_chat()
+    request = make_request(modalities=["text", "audio"])
 
     async def result_generator():
         # Engine only produces text, no audio output at all
-        yield _make_text_omni_output(text="hi", token_ids=[10], finish_reason=None)
-        yield _make_text_omni_output(text="!", token_ids=[11], finish_reason="stop")
+        yield make_text_omni_output(text="hi", token_ids=[10], finish_reason=None)
+        yield make_text_omni_output(text="!", token_ids=[11], finish_reason="stop")
 
     raw_lines = await _collect_stream(
         serving_chat.chat_completion_stream_generator(
@@ -373,12 +243,12 @@ async def test_declared_modality_not_produced_emits_fallback_stop():
 async def test_declared_modality_not_produced_text_finish_suppressed():
     """When text finishes but audio (declared in modalities) never appears,
     the text finish chunk has finish_reason=null (suppressed)."""
-    serving_chat = _build_serving_chat()
-    request = _make_request(modalities=["text", "audio"])
+    serving_chat = build_serving_chat()
+    request = make_request(modalities=["text", "audio"])
 
     async def result_generator():
-        yield _make_text_omni_output(text="hi", token_ids=[10], finish_reason=None)
-        yield _make_text_omni_output(text="!", token_ids=[11], finish_reason="stop")
+        yield make_text_omni_output(text="hi", token_ids=[10], finish_reason=None)
+        yield make_text_omni_output(text="!", token_ids=[11], finish_reason="stop")
         # No audio output — stream ends
 
     raw_lines = await _collect_stream(
